Load all bounding boxes from a streamed model
