+++ src/fragments/FragmentIfcStreamer/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./streaming-settings";
+export * from "./fragment-ifc-stream-converter";
+export * from "./fragment-stream-loader";
 export * from "./base-types";
